+++ src/nex/fmt.clj
@@ -47,6 +47,8 @@
     (case (:type expr)
       :integer (str (:value expr))
       :byte (str (:value expr) "u8")
+      :int16 (str (:value expr) "i16")
+      :int32 (str (:value expr) "i32")
       :real (str (:value expr))
       :string (str "\"" (:value expr) "\"")
       :char (str "'" (:value expr) "'")